backend/requirements.txt
DJango==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
channels==4.1.0
daphne==4.1.2
channels-redis==4.2.0
django-jazzmin==3.0.0
django-cors-headers==4.4.0
pusher==3.3.2
python-dotenv==1.0.1
pywebpush==2.0.0
pyotp==2.9.0
